Add multiple Queue commands: * Set texture * Set render target * Set render state * Set window (viewport) size * Set vertex FVF Fix VS2010 build Partially reimplement RenderScreen in dxglDirectDrawSurface.cpp

+++ ddraw/util.c
@@ -214,8 +214,8 @@ void ShrinkMip(DWORD *x, DWORD *y, int level)
 	int i;
 	for (i = 0; i < level; i++)
 	{
-		*x = max(1, (DWORD)floorf((float)*x / 2.0f));
-		*y = max(1, (DWORD)floorf((float)*y / 2.0f));
+		*x = max(1, (DWORD)floor((float)*x / 2.0f));
+		*y = max(1, (DWORD)floor((float)*y / 2.0f));
 	}
 }
